Alignment refactor
